# node/express-prisma-typescript/src/domains/user/controller/user.controller.ts
import { Request, Response, Router } from 'express'
import HttpStatus from 'http-status'
// express-async-errors is a module that handles async errors in express, don't forget import it in your new controllers
import 'express-async-errors'

import { db } from '@utils'

import { UserRepositoryImpl } from '../repository'
import { UserService, UserServiceImpl } from '../service'

export const userRouter = Router()

// Use dependency injection
const service: UserService = new UserServiceImpl(new UserRepositoryImpl(db))

userRouter.get('/', async (req: Request, res: Response) => {
  const { userId } = res.locals.context
  const { limit, skip } = req.query as Record<string, string>

  const users = await service.getUserRecommendations(userId, { limit: Number(limit), skip: Number(skip) })

  return res.status(HttpStatus.OK).json(users)
})

userRouter.get('/me', async (req: Request, res: Response) => {
  const { userId } = res.locals.context

  const user = await service.getUser(userId)

  return res.status(HttpStatus.OK).json(user)
})

userRouter.get('/:userId', async (req: Request, res: Response) => {
  const { userId: otherUserId } = req.params

  const user = await service.getUser(otherUserId)

  return res.status(HttpStatus.OK).json(user)
})

userRouter.delete('/', async (req: Request, res: Response) => {
  const { userId } = res.locals.context

  const deleted = await service.deleteUser(userId)

  return res.status(HttpStatus.OK).json(deleted)
})

userRouter.post('/upload_profile_image', async (req: Request, res: Response) => {
  const { userId } = res.locals.context

  const picUrl = await service.uploadProfilePic(userId)

  return res.status(HttpStatus.OK).json(picUrl)
})

userRouter.get('/privacy/:user_id', async (req: Request, res: Response) => {
  const userId = req.params.user_id

  const privacy = await service.getPrivacy(userId)

  return res.status(HttpStatus.OK).json(privacy)
})

userRouter.get('/by_username/:username', async (req: Request, res: Response) => {
  const { username } = req.params
  const { limit, before, after } = req.query as Record<string, string>

  const user = await service.getUserByUsername(username, { limit: Number(limit), before, after })

  return res.status(HttpStatus.OK).json(user)
})